huggingface/third-party-models/depth-estimation/Depth-Anything-3/requirements.txt
optimum-rbln
addict==2.4.0
omegaconf>=2.3.0
moviepy==1.0.3
einops==0.8.1
opencv-python>=4.9.0.80
plyfile>=1.1.3
matplotlib>=3.8.0
pycolmap>=3.13.0
trimesh>=4.11.2
evo>=1.34.3
